working on report
