+++ log/2018-11-12_22-31-51_v2_best_hyper_parameter_epoch_20/eval.json
@@ -1,0 +1,23 @@
+{
+    "ENSEMBLE_MIOU": [
+        0.8331351280212402,
+        0.8331351280212402,
+        0.9400942325592041
+    ],
+    "ENSEMBLE_WEIGHT": [
+        [
+            0.4,
+            0.6
+        ],
+        [
+            0.5,
+            0.5
+        ],
+        [
+            0.6,
+            0.4
+        ]
+    ],
+    "EVAL_STEPS": 1024,
+    "FINAL_MIOU": 0.9153795838356018
+}
