+++ log/2018-11-13_16-29-55_v2_best_hyper_parameter_epoch_200_continue/eval.json
@@ -1,0 +1,23 @@
+{
+    "ENSEMBLE_MIOU": [
+        0.8331351280212402,
+        0.8331351280212402,
+        0.9308294057846069
+    ],
+    "ENSEMBLE_WEIGHT": [
+        [
+            0.4,
+            0.6
+        ],
+        [
+            0.5,
+            0.5
+        ],
+        [
+            0.6,
+            0.4
+        ]
+    ],
+    "EVAL_STEPS": 1024,
+    "FINAL_MIOU": 0.9081586599349976
+}
